Separate query-related methods from jobs
