+++ google-cloud-bigquery/src/main/scala/akka/stream/alpakka/googlecloud/bigquery/scaladsl/BigQuery.scala
@@ -15,6 +15,7 @@ object BigQuery
     extends BigQueryRest
     with BigQueryDatasets
     with BigQueryJobs
+    with BigQueryQueries
     with BigQueryTables
     with BigQueryTableData {
